pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pimage-local"
version = "0.1.0"
description = "Raspberry Pi camera app"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "PyYAML",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
files = ["pimage"]
